fix: label should also be deleted when deleting connector

+++ packages/blocks/src/root-block/edgeless/utils/crud.ts
@@ -26,7 +26,7 @@ export function deleteElements(
       if (children.length > 1) {
         surface.doc.deleteBlock(element);
       }
-    } else {
+    } else if (element.id) {
       service.removeElement(element.id);
     }
   });
